monitoring/alert-rules: improve filesystem_full alerts
This reduces the threshold for the filesystem_full alert from 90%
to 80%.
The 7d filesystem_full prediction is now based on 2d to keep a
balance between having alerts without much latency and smoothing
out short IO spikes.
Testwise a new filesystem_full prediction for 30d was added, which
is also based on the previous 30d. This interval should be long enough
for interpolation, to avoid long lasting false positives caused by
short IO spikes.
"borgfs" is now included in the device filter.