modules/simplesamlphp: don't use an empty string as cookie default
When I initially deployed simplesamlphp to my server with leaving
`phpSessionCookieName` and `authTokenCookieName` empty (i.e. unspecified),
I got the following slightly confusing error in my logs when trying to
authenticate as `admin`:

    ERR [d5422ef7b1] Cannot set authentication token cookie: Headers already sent

The reason is actually kinda weird: SimpleSAMLPHP tries to set a cookie
using PHP's `setcookie()`-function[1]. The docs state that `false` is
returned when headers are already sent.

However this turns out to be WRONG. If the cookie name is empty and thus
an empty string is passed to `setcookie()`, `false` is returned as well.
However SimpleSAMLPHP wrongly assumes that this is due to already sent
headers and throws this bogus error message.

[1] https://www.php.net/manual/de/function.setcookie.php