simplesamlphp: 1.17.2 -> 1.19.1
The full change-logs can be found on the vendor's homepage[1]. The
reason for the update is that there were a few minor security advisories
disclosed that affect our current version[2]. None of them seemed easily
exploitable, but since this is a rather sensitive piece of software, an
update is still a good idea.

I successfully deployed this to my own systems and tested a login via
`u2f` and `yubico-otp` for Nextcloud and `wiki-js`. No adjustment of my
existing configuration was necessary for this.

I left out an update of the module for PrivacyIDEA[3] on purpose because
I encountered some issues with `u2f`-auth then. I already filed a bug-report
for this[4].

[1] https://simplesamlphp.org/docs/stable/simplesamlphp-changelog
[2] https://simplesamlphp.org/security
[3] https://github.com/privacyidea/simplesamlphp-module-privacyidea/releases/tag/v1.9
[4] https://github.com/privacyidea/simplesamlphp-module-privacyidea/issues/108