modules/monitoring: disable alert for conntrack collector
In Linux 5.15.65 the following change was introduced:

> netfilter: conntrack: NF_CONNTRACK_PROCFS should no longer default to y
>
> NF_CONNTRACK_PROCFS was marked obsolete in commit 54b07dca68557b09
> ("netfilter: provide config option to disable ancient procfs parts") in
> v3.3.

This means that stats from `/proc/net/stat/nf_conntrack` aren't
available anymore and thus `node_scrape_collector_success` is `0` for
conntrack collectors.

I still decided against disabling the entire conntrack collector because
it also fetches the current size of the conntrack table and the maximum
size from `/sys/net/netfilter/nf_conntrack_{count,max}` and thus it's
still useful for the `node_conntrack_limit`-alert.

I also filed an upstream bug for that[1].

So turning off the alert for the collector is the best solution to avoid
spamming `#admin-alerts` with unactionable notifications.

[1] https://github.com/prometheus/node_exporter/issues/2491