simplesamlphp: 1.19.6 -> 1.19.8, module-privacyidea: 2.1.3 -> 3.0.0
Support for simplesamlphp-2.0 is not ready yet, so we stick with 1.19.
Updating is still relevant though because 1.19.8 is marked as security
release[1] as it updates the Symfony http cache dependency to fix
CVE-2022-24894. Additionally, 1.19.7 also contains a bunch of bugfixes.

While at it, I decided to update the privacyidea integration to the
latest version available (3.0). This contains a few config changes (as
outlined in the option's example of the NixOS module). Already tested
the change on my private infrastructure and the auth flows for Nextcloud
and Meshcentral work fine there.

[1] https://github.com/simplesamlphp/simplesamlphp/releases/tag/v1.19.8